move parser functions into new file, add nested tuple parsing support
